+++ tests/src/test/java/com/datastax/oss/kaap/tests/helm/Pulsar2TlsTest.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright DataStax, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.datastax.oss.kaap.tests.helm;
+
+import lombok.extern.slf4j.Slf4j;
+import org.testng.annotations.Test;
+
+@Slf4j
+@Test(groups = "helm-tls")
+public class Pulsar2TlsTest extends TlsTest {
+
+    @Test
+    public void testPerComponents() throws Exception {
+        test(true, false);
+    }
+
+    @Test
+    public void testGlobal() throws Exception {
+        test(false, false);
+
+    }
+}
